- Corrected package name from `architeture` to `architecture`. - Updated JPA imports from `javax.persistence` to `jakarta.persistence`. - Refactored dependency injection: switched from field-based to constructor-based injection in `ProdutoService` and `ProdutoController`. - Adjusted ArchUnit custom rules for better validation coverage. - Updated `pom.xml`: upgraded Spring Boot to 3.4.6 and JDK to 24, refined dependencies.
