+++ pom.xml
@@ -6,7 +6,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.7.18</version>
+		<version>3.4.6</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 
@@ -17,7 +17,7 @@
 	<description>Example application to demonstrate architecture tests with ArchUnit</description>
 
 	<properties>
-		<java.version>1.8</java.version>
+		<java.version>24</java.version>
 	</properties>
 
 	<dependencies>
@@ -55,21 +55,22 @@
 		<dependency>
 			<groupId>com.tngtech.archunit</groupId>
 			<artifactId>archunit</artifactId>
-			<version>1.0.1</version>
+			<version>1.4.1</version>
 			<scope>test</scope>
 		</dependency>
 		<dependency>
 			<groupId>com.tngtech.archunit</groupId>
-			<artifactId>archunit-junit5-api</artifactId>
-			<version>1.0.1</version>
+			<artifactId>archunit-junit5</artifactId>
+			<version>1.2.2</version>
 			<scope>test</scope>
 		</dependency>
 		<dependency>
-			<groupId>com.tngtech.archunit</groupId>
-			<artifactId>archunit-junit5-engine</artifactId>
-			<version>1.0.1</version>
+			<groupId>org.junit.jupiter</groupId>
+			<artifactId>junit-jupiter</artifactId>
+			<version>5.10.2</version>
 			<scope>test</scope>
 		</dependency>
+
 	</dependencies>
 
 	<build>
@@ -78,18 +79,21 @@
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
 			</plugin>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-surefire-plugin</artifactId>
-				<version>2.22.2</version><!-- Ensure this version is compatible with JUnit 5 Platform -->
-				<dependencies>
-					<dependency>
-						<groupId>org.junit.platform</groupId>
-						<artifactId>junit-platform-surefire-provider</artifactId>
-						<version>1.3.2</version><!-- Or a newer compatible version -->
-					</dependency>
-				</dependencies>
-			</plugin>
+<!--			<plugin>-->
+<!--				<groupId>org.apache.maven.plugins</groupId>-->
+<!--				<artifactId>maven-surefire-plugin</artifactId>-->
+<!--				<version>3.5.3</version>-->
+<!--				<configuration>-->
+<!--					<includes>-->
+<!--						<include>**/*Test.java</include>-->
+<!--						<include>**/*Tests.java</include>-->
+<!--						<include>**/*ArchTest.java</include>-->
+<!--					</includes>-->
+<!--					<systemPropertyVariables>-->
+<!--						<java.util.logging.manager>org.apache.logging.log4j.jul.LogManager</java.util.logging.manager>-->
+<!--					</systemPropertyVariables>-->
+<!--				</configuration>-->
+<!--			</plugin>-->
 		</plugins>
 	</build>
