+++ src/test/java/com/archunit/example/architecture/LayerArchTest.java
@@ -1,4 +1,4 @@
-package com.archunit.example.architeture;
+package com.archunit.example.architecture;
 
 import com.tngtech.archunit.junit.AnalyzeClasses;
 import com.tngtech.archunit.junit.ArchTest;
@@ -12,15 +12,12 @@ public class LayerArchTest {
     public static final ArchRule layers_access_check =
         Architectures.layeredArchitecture()
                 .consideringAllDependencies()
-                .layer("Config").definedBy("com.archunit.example.config..")
                 .layer("Controller").definedBy("com.archunit.example.controller..")
                 .layer("Service").definedBy("com.archunit.example.service..")
                 .layer("Repository").definedBy("com.archunit.example.model.repository..")
                 .layer("Model").definedBy("com.archunit.example.model..", "com.archunit.example.model.domain..")
 
-                // Add back original, more complex access rules
-                .whereLayer("Controller").mayOnlyBeAccessedByLayers("Config")
-                .whereLayer("Service").mayOnlyBeAccessedByLayers("Controller", "Service", "Config")
+                .whereLayer("Service").mayOnlyBeAccessedByLayers("Controller", "Service")
                 .whereLayer("Repository").mayOnlyBeAccessedByLayers("Service")
-                .whereLayer("Model").mayOnlyBeAccessedByLayers("Repository", "Service", "Controller", "Config", "java..");
+                .whereLayer("Model").mayOnlyBeAccessedByLayers("Repository", "Service", "Controller");
 }
